better md5 change detection

+++ jobs/shapefile.sql
@@ -3,6 +3,8 @@ $$
 DECLARE
 	test_count integer;
 	test_mincount integer = 180;
+	test_md5_current text;
+  test_md5_previous text;
 
 BEGIN
   -- create server connection
@@ -32,18 +34,37 @@ BEGIN
   FROM
     ogr_fdw_shapefile_polling_locations;
 
+	SELECT
+    md5(CAST((array_agg(t.*)) AS text)) into test_md5_current
+  FROM
+    (
+      SELECT name, precno, geom
+      FROM ogr_fdw_shapefile_polling_locations
+      ORDER BY 1
+    ) AS t;
+
+	SELECT obj_description('polling_locations'::regclass) into test_md5_previous;
+
 
   -- run update if tests passed
   IF test_count >= test_mincount THEN
-    TRUNCATE TABLE polling_locations RESTART IDENTITY;
-    INSERT INTO polling_locations
-				(geom, precno, name, city, zone, address, type, location)
-			SELECT
-				geom, precno, name, city, zone, address, type, location
-			FROM
-				ogr_fdw_shapefile_polling_locations;
+		IF test_md5_previous <> test_md5_current THEN
+			TRUNCATE TABLE polling_locations RESTART IDENTITY;
+
+			INSERT INTO polling_locations
+					(geom, precno, name, city, zone, address, type, location)
+				SELECT
+					geom, precno, name, city, zone, address, type, location
+				FROM
+					ogr_fdw_shapefile_polling_locations;
+
+
+			EXECUTE 'COMMENT ON TABLE polling_locations IS ' || quote_literal(test_md5_current);
+		ELSE
+			raise exception 'no change found';
+		END IF;
 	ELSE
-		raise exception 'minimum row count failure';
+		raise exception 'error: minimum row count failure';
 	END IF;
 
   -- remove server and foreign table
